Update in diagnostic
I have edited the LCFS analysis as it was being done just outside of the LCFS, so there was a discontinuity at eta=0. I have also corrected the scale of the partial flux surface average of currents, which where out of scale, as I had forgotten to copy part_t1d into part_fsa1d in the else condition in line 450.